Crate wafel_app

The executable for the main Wafel GUI.

It is possible to hot reload the [wafel_app_ui] crate while this binary is running if the reload feature is enabled. Commands to run (in separate terminals):

cargo run -p wafel_app --features reload
cargo watch -w wafel_app_ui -x "build -p wafel_app_ui"

Modules

app 🔒

egui_state 🔒

env 🔒

Environment for the Wafel application.

hot_reload 🔒

logging 🔒

window 🔒

Sets up the main application window and event loop.

Constants

WAFEL_VERSION 🔒

Functions

main 🔒