wafel_app::hot_reload::hot_wafel_app_ui
Trait Env

pub trait Env {
    // Required methods
    fn wafel_version(&self) -> &str;
    fn log_file_path(&self) -> &Path;
    fn processes(&self) -> Vec<ProcessInfo>;
    fn is_process_open(&self, pid: u32) -> bool;
    fn take_recent_panic_details(&self) -> Option<String>;
}

A trait defining the Wafel application’s interaction with the file system and windowing system.

The GUI implementation is in the wafel_app crate, but it can be overridden to run Wafel in headless mode for example.

Required Methods

fn wafel_version(&self) -> &str

Return the current version of Wafel.

fn log_file_path(&self) -> &Path

Return the path to the log file.

fn processes(&self) -> Vec<ProcessInfo>

Return a list of running processes.

fn is_process_open(&self, pid: u32) -> bool

Return true if a process with the given pid is open.

fn take_recent_panic_details(&self) -> Option<String>

Return details of the most recent panic caught by the panic handler.

This method also clears the panic details.